data/requirements.txt
llama-cpp-python==0.2.60
faiss-cpu==1.7.4
langchain==0.1.14
langchain-community==0.0.34
pypdf==4.1.0
tqdm==4.66.1
numpy==1.26.4


